refactor(runs): extract RunDetailPanel component (#1505)
Splits the 666-line RunDetailPage into a reusable RunDetailPanel component
so the same run-detail surface can be embedded inside Lens RunBubble in a
follow-up PR (#1506) without duplicating the summary/steps/ai-trace/files/
approvals/cost implementation.

Behavior identical to main. New `embedded` prop hides breadcrumbs, the
Refresh/Export/Stop header actions, and the outer maxWidth wrapper — all
gated so the current canvas page renders pixel-identically.

- apps/web/src/components/runs/RunDetailPanel.tsx: new file, all previous
  RunDetailPage content lives here; reads workflowId/runId from props.
- apps/web/src/app/(app)/workflows/[id]/runs/[run_id]/page.tsx: shrinks
  from 666 lines to a thin AppShell + RunDetailPanel shell.
- Drops the unused Pill helper (0 call sites).

TypeScript clean. ESLint: 1 warning, identical to the one already on main
(react-hooks/exhaustive-deps on the poll-clearing useEffect) — not a
regression, just moved with the code.
